# OpenAPI.yaml
openapi: "3.1.0"
info:
  title: "Asset_Management API"
  description: "Asset_Management API"
  version: "1.0.0"
servers:
  - url: "https://Asset_Management"
paths:
...
  /assets/{assetUUID}:
    get:
      summary: "GET assets/{assetUUID}"
      operationId: "getById"
      parameters:
        - name: "assetUUID"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssetDTO"
    put:
      summary: "PUT assets/{assetUUID}"
      operationId: "update"
      parameters:
        - name: "assetUUID"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssetDTO"
    delete:
      summary: "DELETE assets/{assetUUID}"
      operationId: "delete"
      parameters:
        - name: "assetUUID"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "204":
          description: "No Content"
...
components:
  schemas:
    AssetDTO:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
        type:
          type: "string"
...